Added classes for items
I added a class for the items that will be added to the document.
The html document is initialised without any items, but when they get
added through javascript, this new class will be applied to them.
Slightly tweaked the items container class to ensure that 4 items
can fit in a row. Added buttons