Added full page functionality
I added button functionality for the clear and filter buttons, and
wrote all the necessary functions for filtering through and displaying
the individual shopping items. Also changed around the 'createDiv'
function to now create the items with an image and description instead
of just the name of the product/theme. I cut some corners and wrote a
ton of unnecessary code, mostly because this was a project aimed at
practicing DOM manipulation and javascript logic. The next prototype for
'CC_photography' will use more objects and effecient code to do filtering
and displaying in order to ensure good practice is followed and the code
is readable.